Move airvisual base entity to separate module
